pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-dhara_ai"
version = "0.1.0"
description = "Dhara AI — Legal contract review and negotiation environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # OpenEnv runtime: brings in FastAPI, Uvicorn, Pydantic, Gradio, websockets.
    "openenv-core[core]>=0.2.2",
    # OpenAI SDK is required only by inference.py (the baseline agent),
    # but we include it so `uv run inference.py` works without a separate install.
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.27.0",        # required by fastapi.testclient
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m dhara_ai.server.app
server = "dhara_ai.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["dhara_ai", "dhara_ai.server"]
package-dir = { "dhara_ai" = ".", "dhara_ai.server" = "server" }